+++ gtk/src/menu.ui
@@ -40,13 +40,17 @@
       <attribute name="label" translatable="yes">_Queue</attribute>
       <section>
         <item>
-          <attribute name="action">app.queue-add</attribute>
+          <attribute name="action">app.add-current</attribute>
           <attribute name="label" translatable="yes">_Add</attribute>
           <attribute name="accel">&lt;alt&gt;A</attribute>
         </item>
         <item>
-          <attribute name="action">app.queue-add-all</attribute>
+          <attribute name="action">app.add-multiple</attribute>
           <attribute name="label" translatable="yes">Add _Multiple</attribute>
+        </item>
+        <item>
+          <attribute name="action">app.add-all</attribute>
+          <attribute name="label" translatable="yes">Add A_ll</attribute>
           <attribute name="accel">&lt;shift&gt;&lt;alt&gt;A</attribute>
         </item>
       </section>
