+++ gtk/src/queuehandler.h
@@ -15,7 +15,7 @@
  * See the GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with callbacks.h.  If not, write to:
+ * along with queuehandler.h.  If not, write to:
  *  The Free Software Foundation, Inc.,
  *  51 Franklin Street, Fifth Floor
  *  Boston, MA  02110-1301, USA.
@@ -31,7 +31,6 @@
 void     ghb_queue_buttons_grey(signal_user_data_t *ud);
 gboolean ghb_reload_queue(signal_user_data_t *ud);
 void     ghb_queue_remove_row(signal_user_data_t *ud, int row);
-void     ghb_finalize_job(GhbValue *settings);
 gint     ghb_find_queue_job(GhbValue *queue, gint unique_id, GhbValue **job);
 void     ghb_low_disk_check(signal_user_data_t *ud);
 void     ghb_queue_progress_set_visible(signal_user_data_t *ud, int index,
@@ -44,5 +43,7 @@ void     ghb_queue_select_log(signal_user_data_t * ud);
 void     ghb_queue_update_live_stats(signal_user_data_t * ud, int index,
                                      ghb_instance_status_t * status);
 void     ghb_queue_drag_n_drop_init(signal_user_data_t * ud);
+void     ghb_add_to_queue_list(signal_user_data_t *ud, GhbValue *queueDict);
+void     ghb_queue_selection_init(signal_user_data_t * ud);
 
 #endif // _QUEUEHANDLER_H_
